pyproject.toml
[tool.poetry]
name = "telegram-media-bot"
version = "1.0.0"
description = "TikTok/Instagram Media Downloader Bot"
author = "Bot Developer"

[tool.poetry.dependencies]
python = "^3.8"
yt-dlp = "^2024.1.0"
python-telegram-bot = "^20.0"
instaloader = "^4.10"
aiohttp = "^3.8.0"
aiofiles = "^23.0.0"
python-dotenv = "^0.19.0"
pyyaml = "^6.0"
pillow = "^9.0.0"
requests = "^2.28.0"
flask = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"